IncomeCategory,Income,Expense,ExpenseCategory layers done

+++ src/main/java/project/expenseincomeproject/service/IncomeService.java
@@ -1,5 +1,6 @@
 package project.expenseincomeproject.service;
 
+import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import project.expenseincomeproject.dto.IncomeRequestDto;
@@ -15,6 +16,7 @@ import project.expenseincomeproject.repository.UserRepository;
 
 import java.time.LocalDate;
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -44,6 +46,7 @@ public class IncomeService {
                 .orElseThrow(() -> new RuntimeException("Income not found with id: " + id));
 
         // IncomeRequestDto-dan məlumatları alıb mövcud gəliri yeniləmək
+        existingIncome.setName(incomeRequestDto.getName());
         existingIncome.setAmount(incomeRequestDto.getAmount());
         existingIncome.setDate(incomeRequestDto.getDate());
 
@@ -66,19 +69,39 @@ public class IncomeService {
 
 
     public void deleteIncome(Long id) {
+        Optional<Income> income = incomeRepository.findById(id);
+        if (income.isEmpty()) {
+            throw new EntityNotFoundException("Income not found with id: " + id);
+        }
         incomeRepository.deleteById(id);
     }
 
-    public List<IncomeResponseDto> getIncomesByCategoryAndDateRange(Long categoryId, LocalDate startDate, LocalDate endDate) {
-        List<Income> incomes = incomeRepository.findByIncomeCategoryIdAndDateBetween(categoryId, startDate, endDate);
+    public List<IncomeResponseDto> getIncomesByCategoryNameAndDateRange(String categoryName, LocalDate startDate, LocalDate endDate) {
+
+        // Əgər belə bir categoryName bazada yoxdursa, istisna at
+        if (!incomeCategoryRepository.existsByIncomeCategoryName(categoryName)) {
+            throw new RuntimeException("Category not found!");
+        }
+
+        // Kateqoriya varsa, indi gəlirləri axtaraq
+        List<Income> incomes = incomeRepository.findByIncomeCategory_IncomeCategoryNameAndDateBetween(categoryName, startDate, endDate);
+
+        // Əgər bu kateqoriyaya aid gəlirlər yoxdursa, başqa bir istisna atırıq
+        if (incomes.isEmpty()) {
+            throw new RuntimeException("Resource not found!");
+        }
+
         return incomes.stream()
                 .map(incomeMapper::toIncomeResponseDto)
                 .collect(Collectors.toList());
     }
 
-
     public List<IncomeResponseDto> getIncomesByDateRange(LocalDate startDate, LocalDate endDate) {
         List<Income> incomes = incomeRepository.findByDateBetween(startDate, endDate);
+        if (incomes.isEmpty()) {
+            throw new RuntimeException("Resource not found !");
+        }
+
         return incomes.stream()
                 .map(incomeMapper::toIncomeResponseDto)
                 .collect(Collectors.toList());
